+++ src/app/app.ts
@@ -3,9 +3,10 @@
  *
  * @module
  */
+
+import * as sskts from '@motionpicture/sskts-domain';
 import * as bodyParser from 'body-parser';
 import * as express from 'express';
-import * as mongoose from 'mongoose';
 
 import basicAuth from './middlewares/basicAuth';
 import errorHandler from './middlewares/errorHandler';
@@ -16,8 +17,8 @@ const app = express();
 app.use(basicAuth); // ベーシック認証
 
 // view engine setup
-app.set('views', `${__dirname}/views`);
-app.set('view engine', 'ejs');
+// app.set('views', `${__dirname}/views`);
+// app.set('view engine', 'ejs');
 
 app.use(bodyParser.json());
 // The extended option allows to choose between parsing the URL-encoded data
@@ -28,16 +29,13 @@ app.use(bodyParser.urlencoded({ extended: true }));
 // app.use(express.static(__dirname + '/../public'));
 
 // mongoose
-(<any>mongoose).Promise = global.Promise;
-mongoose.connect(process.env.MONGOLAB_URI);
+sskts.mongoose.connect(<string>process.env.MONGOLAB_URI);
 
 // routers
-import clientEventsRouter from './routers/clientEvents';
 import gmoRouter from './routers/gmo';
 import router from './routers/router';
 import sendgridRouter from './routers/sendgrid';
 app.use('/', router);
-app.use('/clientEvents', clientEventsRouter);
 app.use('/gmo', gmoRouter);
 app.use('/sendgrid', sendgridRouter);