+++ src/gmo.spec.ts
@@ -1,26 +1,16 @@
 /**
  * GMOルーターテスト
- *
  * @ignore
  */
-import * as sskts from '@motionpicture/sskts-domain';
+
 import * as assert from 'assert';
 import * as fs from 'fs-extra';
 import * as HTTPStatus from 'http-status';
-import * as mongoose from 'mongoose';
 import * as supertest from 'supertest';
 
-import * as app from '../app/app';
+import * as app from './app/app';
 
 describe('GMO結果通知', () => {
-    let connection: mongoose.Connection;
-    before(async () => {
-        // 全て削除してからテスト開始
-        connection = mongoose.createConnection(process.env.MONGOLAB_URI);
-        const gmoNotificationAdapter = sskts.adapter.gmoNotification(connection);
-        await gmoNotificationAdapter.gmoNotificationModel.remove({}).exec();
-    });
-
     it('不正なリクエスト', async () => {
         await supertest(app)
             .post('/gmo/notify')
@@ -43,11 +33,6 @@ describe('GMO結果通知', () => {
             .expect(HTTPStatus.OK)
             .then(async (response) => {
                 assert.equal(response.text, '0');
-                const gmoNotificationAdapter = sskts.adapter.gmoNotification(connection);
-                const notificationDoc = await gmoNotificationAdapter.gmoNotificationModel.findOne(
-                    { order_id: notification.OrderID }
-                ).exec();
-                assert.notEqual(notificationDoc, null);
             });
     });
 });